French employment agreement ; LibLicense model form
The FR employment agreement leaves many blanks, and presents issues of
cross-language inclusion.  I can see solutions, but need to work
through.  In the meantime, the substance - the employment agreement
text, is available to be codified.  LibLicense is a pretty heavy-duty
copyright license for use in obtaining access to a publisher’s
collection.