Deal_Room:  Galion: SeriesSeed:  Investment Agreement
Added the Series Seed Investment Agreement (with one missing point) to
the “Seed” round.  This is in order to deepen the example and get more
sense of the naming schemes and challenges.

Mostly, it went smoothly and the practice of including the form with a
prefix of the rounds seems to work.  Probably it makes sense to extend
this practice of prefixing to an even greater extent - to prefix the
Term Sheet with e.g.,  Seed.TS.=[Wx/com/…].  The de-prefixing would
mean that this made very little (or no) difference vis-a-vis prefixing
with Seed.=[Wx/com/…] and would allow us to put customizing terms
anywhere in the Dashboards, giving us full control over how we organize
our info.