IACCM Cooperate - BIG reworking - Outline/Sec
IACCM-Cooperate-CmA:  The goal is to have the “Outline” consolidate the
structure.  The Sec/
-DocType to have the list of Sections.  And - big change - to simplify
the internal cross-references within a document - make them generic -
so that a reference to Conf.Xref will always work, no matter how the
document is organized, and no matter what choice of Sec/Conf/… the
author makes.